% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distribution_shift.R
\name{analyze_distribution_dynamics}
\alias{analyze_distribution_dynamics}
\title{Calculate distribution shift metrics}
\usage{
analyze_distribution_dynamics(
  r_stack,
  trailing_p = 0.05,
  leading_p = 0.95,
  core_p = 0.5,
  all_p = 0.95,
  poly_dir = NULL
)
}
\arguments{
\item{r_stack}{a spatraster of sdm results}

\item{trailing_p, leading_p}{probabilities for leading and trailing edges; defaults to 5\% and 95\%}

\item{core_p, range_p}{probabilities associated with desired core and range KDEs; defaults to 50\% and 95\% respectively}

\item{poly_dir}{location to save kde polygons}
}
\value{
a list containing a data.frame of metrics where the number of rows is equal to the number of timestamps, the core kde polygons, and the range kde polygons
}
\description{
calculate trailing and leading edges of species distributions, desired kdes, kde areas, and weighted centroids
}
