% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_logic_rule.R
\name{apply_logic_rule}
\alias{apply_logic_rule}
\title{Calculate Total Sensitivity Score with the Logic Rule}
\usage{
apply_logic_rule(attribute_vec, bootstrap)
}
\arguments{
\item{attribute_vec}{output from \code{attribute_score}}

\item{bootstrap}{binary TRUE/FALSE to turn on/off bootstrapping of sensitivity scores. Will impact outputs.}
}
\value{
returns the final sensitivity scores based on the logic rule. The output is similar to \code{attribute_score}, where if \code{bootstrap = TRUE}, the function returns a vector containing the final sensitivity score for each sample, and if \code{bootstrap = FALSE}, it returns a single final value.
}
\description{
Calculates total sensitivity from expert counts or bootstrapped values. Based on code from the Highly Migratory Species (HMS) CVA team.
}
