% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_env.R
\name{avg_env}
\alias{avg_env}
\title{Average MOM6 Data}
\usage{
avg_env(rawList)
}
\arguments{
\item{rawList}{List of output rasters from \code{pull_hind} or \code{pull_forecast}}
}
\value{
a list whose length is equal to the number of variables supplied, where each item in the list is a rasterStack of data associated with that variable
}
\description{
Calculate monthly average from raw MOM6 data across the entire provided timeseries
}
