% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_model_data.R
\name{avg_model_data}
\alias{avg_model_data}
\title{Average Model Data}
\usage{
avg_model_data(raw, spatial_temporal)
}
\arguments{
\item{raw}{spatRaster from \code{pull_mom6_hindcast} or \code{pull_mom6_forecast}}

\item{spatial_temporal}{TRUE/FALSE to determine averaging method. If TRUE, averages are spatially and temporally explicit. If FALSE, averaging occurs across space and time}
}
\value{
If \code{spatial.temporal} is TRUE, a spatRaster is returned. If FALSE, a single value is returned.
}
\description{
Calculate monthly or global averages from raw model data across the entire provided timeseries. This is built specifically for MOM6 output, but would work on any spatRaster of gridded data.
}
