% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhatt_coeff_brt.R
\name{bhatt_coeff_brt}
\alias{bhatt_coeff_brt}
\title{Bhattacharyya Coefficient of two distributions}
\source{
From Camrin Brawn (WHOI): https://zenodo.org/records/7971532.
}
\usage{
bhatt_coeff_brt(x, y, bw = bw.nrd0, ...)
}
\arguments{
\item{x}{a numeric vector of length >= 2.}

\item{y}{a numeric vector of length >= 2.}

\item{bw}{can be either a fixed value of bins or a function to calculate the bandwidth of each bin (see ?bw.nrd). Default is bw.nrd0.}

\item{...}{any optional arguments to be passed to the given bw function.}
}
\description{
Calculates the Bhattacharyya Coefficient (BC) between two distributions. BC is the sum of the square root of the multiplication of the elements in bin i for both distributions. It ranges from 0 (no overlap) to 1 (full overlap). v.0.1. Used within \code{eval_brt}. Included to ensure functionality.
}
