% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhattacharyya_stat.R
\name{bhattacharyya_stat}
\alias{bhattacharyya_stat}
\title{Bhattacharyya's coefficient test}
\source{
From Camrin Brawn (WHOI): https://zenodo.org/records/7971532 and http://tguillerme.github.io/R/bhatt.coef.R
}
\usage{
bhattacharyya_stat(data, response, vars)
}
\arguments{
\item{data}{data used to test the model. Must have a presence absence column called "presabs" and column names that match the names in vars}

\item{response}{name of the response variable}

\item{vars}{character vector of the desired variables}
}
\value{
vector of Bhattacharyya's coefficients for each variable
}
\description{
test Bhattacharyya's coefficient ranges from 0 to 1, with 0 being no overlap in distributions and 1 being perfect overlap. Used within \code{eval_brt}. Included to ensure functionality.
}
