% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.certainty.R
\name{bootstrap.certainty}
\alias{bootstrap.certainty}
\title{Calculate Certainty on Bootstrapped Sensitivity Scores}
\usage{
bootstrap.certainty(bootstrap.scores, expert.scores)
}
\arguments{
\item{bootstrap.scores}{output from \code{calculate.sensitivity(bootstrap = TRUE)}}

\item{expert.scores}{output from \code{calculate.sensitivity(bootstrap = FALSE)} to append bootstrap certainty to.}
}
\value{
returns the \code{expert.scores} with a new column called 'Certainty', which includes the percentage of bootstrapped sensitivities that matched the weighted average final sensitivity.
}
\description{
Calculates certainty - the percentage of bootstrapped values that match expert scores - on sensitivity
}
