% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_certainty.R
\name{bootstrap_certainty}
\alias{bootstrap_certainty}
\title{Calculate Certainty on Bootstrapped Sensitivity Scores}
\usage{
bootstrap_certainty(bootstrap_scores, expert_scores)
}
\arguments{
\item{bootstrap_scores}{output from \code{calculate_sensitivity(bootstrap = TRUE)}}

\item{expert_scores}{output from \code{calculate_sensitivity(bootstrap = FALSE)} to append bootstrap certainty to.}
}
\value{
returns the \code{expert_scores} with a new column called 'Certainty', which includes the percentage of bootstrapped sensitivities that matched the weighted average final sensitivity.
}
\description{
Calculates certainty - the percentage of bootstrapped values that match expert scores - on sensitivity
}
