% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ensemble_wrapper.R
\name{build_ensemble_wrapper}
\alias{build_ensemble_wrapper}
\title{Build and Predict Ensemble Model}
\usage{
build_ensemble_wrapper(spp, yr_min, yr_max, build_ens = T)
}
\arguments{
\item{spp}{Species name to add to log files and save data to correct directory (see vignette for recommended directory set up)}

\item{yr_min}{start of year range to specify which predictions to use}

\item{yr_max}{end of year range to specify which predictions to use}

\item{build_ens}{TRUE/FALSE to switch on/off ensemble model building, set to FALSE to just predict ensemble}
}
\value{
returns the file path of the saved predictions. Predicted rasters, model weights, and preds from the ensemble are saved within specific directories. See the manual for recommended directory set up.
}
\description{
A wrapper function to build and predict the ensemble models using \code{build_sdm} and \code{make_sdm_predictions}. Progress is tracked via a log file. NOTE that this function does NOT include skip functionality. This is because it is quicker than the other predictions since it is performing a weighted average of the individual model predictions.
}
