% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fisheries_df.R
\name{build_fisheries_df}
\alias{build_fisheries_df}
\title{Convert Standardized Fisheries Data into Presence/Absence/Effort Data Frame}
\usage{
build_fisheries_df(
  data,
  is_obs,
  grid,
  tm_multiplier = 24 * 60 * 60,
  origin = "1993-01-01",
  all_names
)
}
\arguments{
\item{data}{a data.frame containing the data to use from a specific source}

\item{is_obs}{a TRUE/FALSE indicating whether or not the data source is observer or similar fisheries-dependent data. Will force function to check if species is observed at least 30 times throughout timeseries before creating using data}

\item{grid}{static link to a ncdcf object with the variables lon, lat, time - can be link to remote data - must be able to be read with nc_open}

\item{tm_multiplier}{multiplier to help convert timestep to POSIX (seconds since origin), defaults to 86400 (number of seconds in a day)}

\item{origin}{Origin of time series to be used by POSIXct}

\item{all_names}{a vector containing all possible names for the target species. Must have a length >= 1}
}
\value{
a data.frame with the same extent as the provided grid, and a number of layers equal to the timeseries associated with the provided model data
}
\description{
Takes output from \code{standardize_data} and matches it to the provided grid ranging from 0-2 where:
\itemize{
\item 0 - Fishing effort present for the target species in the cell
\item 1 - Target species caught in the cell
}
}
