% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_preds_df.R
\name{build_preds_df}
\alias{build_preds_df}
\title{Build Prediction Data Frame}
\usage{
build_preds_df(observations, xy_col, prediction_rasters)
}
\arguments{
\item{observations}{a data.frame. A data.frame including fisheries observations.}

\item{xy_col}{a vector with a length of 2 indicating the longitude and latitude column names}

\item{prediction_rasters}{spatRaster of predicted model values on desired timeseries. Values will be matched to observations in \code{data}.}
}
\description{
Match test data to model prediction rasters in a data.frame. Used to calculate AUC in \code{calculate_sdm_auc} and \code{ensemble_sdm_wrapper.R}
}
