% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sdm.R
\name{build_sdm}
\alias{build_sdm}
\title{Build Component or Ensemble Species Distribution Models}
\usage{
build_sdm(
  se,
  pa_col,
  xy_col,
  month_col = "month",
  year_col = "year",
  var_names,
  model,
  year_range = range(se[, year_col], na.rm = T),
  ensemble_weights = NULL,
  ensemble_preds = NULL
)
}
\arguments{
\item{se}{data frame containing species presence/absence data and desired environmental covariate data.}

\item{pa_col}{column name for presence/absence column}

\item{xy_col}{a vector with a length of 2 indicating the longitude and latitude column names}

\item{month_col, year_col}{column names for month and year columns respectively. Defaults to 'month' and 'year' respectively.}

\item{var_names}{a vector of covariate names to use in the desired model. Should match some or all of the column names in \code{se}.}

\item{model}{one of the following indicating the desired model to build: gam, maxent, brt, rf, sdmtmb, or ens}

\item{year_range}{vector with length of two including the maximum and minimum years to include in the model. Only used if \code{model = 'sdmtmb'}. Should include maximum desired year if predicting as well; i.e. if training data range from 1993-2019, but you want to forecast out to 2035, the \code{year_range} should be c(1993,2035). Defaults to the range of the \code{year_col} in \code{se}}

\item{ensemble_weights}{vector of model weights used to build ensemble model. The vector must have the same length and be in the same order as the corresponding predictions list.}

\item{ensemble_preds}{a list of prediction values from component models used to build ensemble model. The list of predictions must have the same length and be in the same order as the corresponding weight vector.}
}
\value{
the model object from the desired model. Model objects will differ depending on the model type.
}
\description{
Build one of the five component models or the final ensemble model
}
