% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExposure.R
\name{calcExposure}
\alias{calcExposure}
\title{Calculate Raw Exposure from Environmental Data}
\usage{
calcExposure(presentRasts, futureRasts)
}
\arguments{
\item{presentRasts, futureRasts}{lists of rasterStacks, representing the present and future timeseries to calculate exposure with, with each item in the list corresponding to an environmental variable. Both lists need to have the same length (aka number of variables) but the lengths of the timeseries (equal to the number of layers in rasterStacks) can differ. Function assumes that the rasters are monthly timeseries and will average monthly.}
}
\value{
a list of rasterStacks with each layer representing monthly raw exposure. The length of the list is equal to the length of the lists supplied as \code{presentRasts} and \code{futureRasts}
}
\description{
Calculates the raw variables exposure as the ratio of the difference between future and present conditions, divided by present standard deviations (a z-score).
}
