% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_directionality.R
\name{calculate_directionality}
\alias{calculate_directionality}
\title{Calculate Directionality Score}
\usage{
calculate_directionality(species, bootstrap = TRUE, samples = 10000)
}
\arguments{
\item{species}{a data frame consisting of all expert directionality scores for a single species.}

\item{bootstrap}{binary TRUE/FALSE to turn on/off bootstrapping of sensitivity scores. Will impact outputs.}

\item{samples}{number of samples to run for bootstrapping. Default is 10,000.}
}
\value{
weighted averaged directionality score. If \code{bootstrap = TRUE}, this is a vector containing bootstrapped scores, with a length equal to the number of samples. If \code{bootstrap = FALSE}, this is a single score.
}
\description{
Calculates weighted average directionality scores for individual sensitivity attributes, with or without bootstrapping. Based on code from the Highly Migratory Species (HMS) CVA team.
}
