% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_model_confidence.R
\name{calculate_model_confidence}
\alias{calculate_model_confidence}
\title{Calculate Model Confidence Score for each Species}
\usage{
calculate_model_confidence(species)
}
\arguments{
\item{species}{a data frame containing all of the expert scores for one species in a column called \code{Score}. See example for how to generate this list from the combined model confidence score spreadsheets.}
}
\value{
a data frame including mean and standard deviation of data quality scores for each attribute with the length of the number of attributes.
}
\description{
Calculates mean and standard deviation of the model confidence score
}
\examples{
\dontrun{
species.data.list <- split(data, data$Species)
#data is the combined data frame of
#all model confidence spreadsheets from scorers

species.conf <- lapply(species.data.list, model.confidence)

speciesMC <- do.call(rbind, species.conf)
#this combines the list of vectors into a data.frame
}

}
