% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sdm_auc.R
\name{calculate_sdm_auc}
\alias{calculate_sdm_auc}
\title{Calculate Model AUC}
\usage{
calculate_sdm_auc(model, data_type, data, prediction_rasters)
}
\arguments{
\item{model}{one of the following indicating the desired model to calculate the metric for from: gam, maxent, brt, rf, sdmtmb, or ens}

\item{data_type}{either 'cv' or 'external'. Determines which methods are used for calculating AUC. If 'cv', then the provided predicted and observed values from a cross-validation are used. If 'external', then predicted values are matched to a provided data.frame of external observations not used in the model construction, which are then used to calculate AUC}

\item{data}{a data.frame. If \code{data_type} is 'cv', then this is the output from \code{pull_sdm_preds}. If \code{data_type} is 'external', then it is a data.frame of observations not used to train the data}

\item{prediction_rasters}{spatRaster of predicted model values on desired timeseries. Values will be matched to observations in \code{data}.}
}
\value{
a value representing the AUC for the given model
}
\description{
Calculate Area Under the ROC Curve (AUC) for one of the component models or the final ensemble from either cross-validation data or on an external dataset
}
