% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sdm_variable_importance.R
\name{calculate_sdm_variable_importance}
\alias{calculate_sdm_variable_importance}
\title{Calculate Variable Importance}
\usage{
calculate_sdm_variable_importance(
  mod,
  se,
  pa_col,
  xy_col,
  month_col = "month",
  year_col = "year",
  var_names,
  model
)
}
\arguments{
\item{mod}{the output from \code{make_sdm} - only used for ensemble}

\item{se}{data frame containing species presence/absence data and desired environmental covariate data.}

\item{pa_col}{column name for presence/absence column}

\item{xy_col}{a vector with a length of 2 indicating the longitude and latitude column names}

\item{month_col, year_col}{column names for month and year columns respectively. Defaults to 'month' and 'year' respectively.}

\item{var_names}{a vector of covariate names to use in the desired model. Should match some or all of the column names in \code{se}.}

\item{model}{one of the following indicating the desired model to calculate variable importance for: gam, maxent, brt, rf, or sdmtmb}
}
\value{
a vector of the variable importance for the given model. Each model calculates these differently, so the values should be normalized in order to compare across models.
}
\description{
Calculate variable importance for one of the component models
}
