% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_in_distribution.R
\name{change_in_distribution}
\alias{change_in_distribution}
\title{Calculate Change in Distribution Metrics}
\usage{
change_in_distribution(abund, area.threshold = 0.75, cell.area = 8 * 8)
}
\arguments{
\item{abund}{model predictions, output from \code{make_predictions} as a rasterStack. Layers need names that correspond to the time stamps represented.}

\item{area.threshold}{a value between 0 and 1, defaults to 0.75. The total area of grid cells with the probability of occurrence of a species equal to or above this threshold will be calculated.}

\item{cell.area}{a number representing the area of a single model grid cell on which the model is predicted.}
}
\value{
a data.frame with three columns: 1) timestamp, equal to the names of the layers in \code{abund}; 2) COG; 3) Area of probabilities greater than the \code{area.threshold}
}
\description{
Calculate two metrics to help describe changes in species distributions - Center of Gravity and Area of Preferred Habitat
}
