% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_exposure_maps.R
\name{combine_exposure_maps}
\alias{combine_exposure_maps}
\title{Make Map of Total Species-Specific Exposure}
\usage{
combine_exposure_maps(variable_maps, count_all, weights, weight_threshold)
}
\arguments{
\item{variable_maps}{rasterStack output from \code{make_exposure_map}}

\item{count_all}{TRUE/FALSE to use \code{weights} and \code{wThreshold} to subset variables to only important variables}

\item{weights}{output from \code{combine_weights} - a vector of variable weights in ensemble SDM}

\item{weight_threshold}{numeric value used to subset weights, variables with weights less to or equal to this value will be excluded from total exposure calculation}
}
\value{
A single raster representing total exposure across space
}
\description{
Uses the logic rule from CVA1.0 to combine variable-specific exposures to create a total exposure map
}
