% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_fisheries_wrapper.R
\name{combine_fisheries_wrapper}
\alias{combine_fisheries_wrapper}
\title{Combine Multiple Rasters of Fisheries Data}
\usage{
combine_fisheries_wrapper(name, skip)
}
\arguments{
\item{name}{Species name to add to log files and save data to correct directory (see vignette for recommended directory set up)}

\item{skip}{TRUE/FALSE indicating whether to skip creating the raster file if file already exists}
}
\value{
\code{combine_rasters_wrapper} returns the range of the rasterBrick returned by \code{merge_rasts}. This should be equal to 0 2, or else there are no presences in the dataset and the models will fail. This function will also save the resulting rasterBrick as a netcdf file in the species' input_rasters folder
}
\description{
A wrapper function for the \code{merge_rasts} function. This function adds log and skip functionality to help with batch runs and running \code{merge_rasts} in parallel.
}
