% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineMaps.R
\name{combineMaps}
\alias{combineMaps}
\title{Make Map of Total Species-Specific Exposure}
\usage{
combineMaps(mapExp, countAll, weights, wThreshold)
}
\arguments{
\item{mapExp}{rasterStack output from \code{makeExposureMap}}

\item{countAll}{TRUE/FALSE to use \code{weights} and \code{wThreshold} to subset variables to only important variables}

\item{weights}{output from \code{combineWeights} - a vector of variable weights in ensemble SDM}

\item{wThreshold}{numeric value used to subset weights, variables with weights less to or equal to this value will be excluded from total exposure calculation}
}
\value{
A single raster representing total exposure across space
}
\description{
Uses the logic rule from CVA1.0 to combine variable-specific exposures to create a total exposure map
}
