% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineWeights.R
\name{combineWeights}
\alias{combineWeights}
\title{Normalize Dynamic Variable Weights}
\usage{
combineWeights(vars, ensWeights, impFlist)
}
\arguments{
\item{vars}{character vector of variables to generate normalized and averaged ensemble importance for}

\item{ensWeights}{vector of component model weights in the ensemble}

\item{impFlist}{character vector of file paths to variable importance for component models}
}
\value{
A vector representing the weighted average of normalized variable importance, representing variable importance in the final ensemble SDM.
}
\description{
Normalizes the component model variable importance, as well as the model weights in the final ensemble, to create a weighted average vector of variable importance for the ensemble model. Options allow for static variables, such as time and space, or non-dynamic environmental variables such as bathymetry, to be excluded from the normalization and weighted average.
}
