% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rast.R
\name{create_rast}
\alias{create_rast}
\title{Convert Standardized Fisheries Data into Presence/Absence/Effort Raster}
\usage{
create_rast(
  data,
  isObs = FALSE,
  grid,
  tmMult = 24 * 60 * 60,
  origin = "1993-01-01",
  targetVec
)
}
\arguments{
\item{data}{data frame to convert to presence/absence raster}

\item{isObs}{TRUE/FALSE indicating whether or not the data is observer or similar fisheries-dependent data. Will force function to check if species is observed at least 30 times throughout timeseries before creating raster}

\item{grid}{static link to a ncdcf object with the variables lon, lat, time - can be link to remote data - must be able to be read with nc_open}

\item{tmMult}{multiplier to help convert timestep to POSIX (seconds since origin), defaults to 86400 (number of seconds in a day)}

\item{origin}{Origin of time series to be used by POSIXct}

\item{targetVec}{a vector containing all possible names for the target species. Must have a length >= 1}
}
\value{
a rasterBrick with the same extent as the provided grid, and a number of layers equal to the timeseries associated with the provided model data
}
\description{
Takes output from \code{standardize_data} and converts it into a raster with the spatial and temporal resolution of the provided grid ranging from 0-2 where:
\itemize{
\item 0 - No fishing effort in the cell
\item 1 - Fishing effort present for the target species in the cell
\item 2 - Target species caught in the cell
}
}
