% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVA2.0-package.R
\docType{package}
\name{CVA2.0-package}
\alias{CVA2.0}
\alias{CVA2.0-package}
\title{CVA2.0: Climate Vulnerability Assessment 2.0 Model Builing and Analysis Methods}
\description{
This package performs all of the necessary analysis for CVA2.0. This includes pulling and formatting the necessary MOM^ and fisheries data; building and predicting component and ensemble models; calculating exposure from model predictions; and calculating and bootstrapping sensitivity from expert scores.
}
\author{
\strong{Maintainer}: Katherine Gallagher \email{katherine.gallagher@noaa.gov}

}
\keyword{internal}
