% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quality.score.R
\name{data.quality.score}
\alias{data.quality.score}
\title{Calculate Data Quality Scores for each Attribute & Species}
\usage{
data.quality.score(species.attributes)
}
\arguments{
\item{species.attributes}{a data frame containing all of the expert scores for one species. See example for how to generate this list from the FCVA output.}
}
\value{
a vector of mean data quality scores for each attribute with the length of the number of attributes.
}
\description{
Calculates mean data quality score
}
\examples{
\dontrun{
data <-read.csv('expert_scores.csv') #sensitivity data from FSCVA Portal
species.data.list <- split(data, data$Stock.Name)

species.dqs <- lapply(species.data.list, data.quality.score)

speciesDQ <- do.call(rbind, species.dqs)
#this combines the list of vectors into a data.frame
}
}
