% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionality_certainity.R
\name{directionality.certainty}
\alias{directionality.certainty}
\title{Calculate Certainty on Bootstrapped Directionality Scores}
\usage{
directionality.certainty(bootstrap.scores, directionality.scores)
}
\arguments{
\item{bootstrap.scores}{output from \code{directionality(bootstrap = TRUE)}}

\item{directionality.scores}{output from \code{directionality(bootstrap = FALSE)} to append bootstrap certainty to.}
}
\value{
returns a vector with a length of two: 1) the output from \code{directionality(bootstrap = FALSE)} and 2) the certainty value, which is the percentage of bootstrapped sensitivities that matched the weighted average final directionality
}
\description{
Calculates certainty - the percentage of bootstrapped values that match expert scores - on directionality
}
