% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionality_certainity.R
\name{directionality_certainty}
\alias{directionality_certainty}
\title{Calculate Certainty on Bootstrapped Directionality Scores}
\usage{
directionality_certainty(bootstrap_scores, directionality_scores)
}
\arguments{
\item{bootstrap_scores}{output from \code{calculate_directionality(bootstrap = TRUE)}}

\item{directionality_scores}{output from \code{calculate_directionality(bootstrap = FALSE)} to append bootstrap certainty to.}
}
\value{
returns a vector with a length of two: 1) the output from \code{calculate_directionality(bootstrap = FALSE)} and 2) the certainty value, which is the percentage of bootstrapped sensitivities that matched the weighted average final directionality
}
\description{
Calculates certainty - the percentage of bootstrapped values that match expert scores - on directionality
}
