% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_brt.R
\name{eval_brt}
\alias{eval_brt}
\title{Calculate BRT model evaluation statistics}
\source{
Much of this code is derived from https://github.com/elhazen/PA-paper
}
\usage{
eval_brt(model, test_data, response, plot = TRUE)
}
\arguments{
\item{model}{a BRT model object}

\item{test_data}{is data used to fit the model. Must contain a column name that matches response.}

\item{response}{is character providing the name of the response variable in test_data}

\item{plot}{TRUE/FALSE to turn on/off plotting}
}
\value{
model evaluation statistics
}
\description{
Calculate evaluation statistics for BRT models. From Camrin Brawn (WHOI): https://zenodo.org/records/7971532. Included to ensure functionality.
}
