% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_kfold_brt.R
\name{eval_kfold_brt}
\alias{eval_kfold_brt}
\title{K-fold BRT fit}
\usage{
eval_kfold_brt(
  data_input,
  gbm_x,
  gbm_y,
  learning_rate = 0.05,
  k_folds = 5,
  tree_complexity = 3,
  bag_fraction = 0.6,
  is_fixed = TRUE,
  max_trees = 2000
)
}
\arguments{
\item{data_input}{input data frame}

\item{gbm_x}{names of predictor variables in \code{data_input}}

\item{gbm_y}{name of response variable in \code{data_input}}

\item{learning_rate}{sets the weight applied to individual trees input to dismo::gbm.step}

\item{k_folds}{number of folds to perform}

\item{tree_complexity}{is tree complexity input to dismo::gbm.step - sets complexity of individual trees}

\item{bag_fraction}{is bag fraction input to dismo::gbm.step - sets the proportion of observations used in selecting variables}

\item{is_fixed}{TRUE/FALSE to determine if gbm.step or gbm.fixed should be used.}

\item{max_trees}{maximum number of trees to fit before stopping}
}
\value{
a list containing 1) the output from \code{eval_brt} and 2) a data.frame of the observed and predicted values used to calculate RMSE/AUC by \code{sdm_cv}
}
\description{
k-fold cross validation code from Camrin Brawn (WHOI): https://zenodo.org/records/7971532. Function originally by B. Abrahms from https://github.com/briana-abrahms/DynamicEnsembleSDM/blob/master/model_evaluation.R. Modified to return information necessary to calculate evaluation metrics. Follows many of the same inputs as \code{gbm.step}
}
\examples{
\dontrun{
eval_kfold_brt(data_input_Fit, gbm_x=c("curl","ild", "ssh", "sst","sst_sd"), "presabs")
}

}
