% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_ensemble.R
\name{evaluate_ensemble}
\alias{evaluate_ensemble}
\title{Test the Ensemble Model with Observations}
\usage{
evaluate_ensemble(spp, spp_names, sources, yr_min, yr_max)
}
\arguments{
\item{spp}{species name. Used to help pull correct ensemble model.}

\item{spp_names}{a vector containing all possible names for the target species. Must have a length >= 1}

\item{sources}{a vector containing all of the source csvs containing the observation data to use in the test. These are loaded in directly by \code{read.csv} so they should either be full file paths or in the current working directory.}

\item{yr_min}{start of year range to specify which predictions to use}

\item{yr_max}{end of year range to specify which predictions to use}
}
\value{
returns the AUC value for the given model on the new data
}
\description{
Calculate Area under the Curve (AUC) for the ensemble species distribution model on an external set of data (not used for building the model). This function builds a dataset of presence/absence with the new data across a variety of sources, extracts the predicted values from the ensemble, and calculates the resulting AUC.
}
