% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_sdm.R
\name{evaluate_sdm}
\alias{evaluate_sdm}
\title{Calculate Performance Metric}
\usage{
evaluate_sdm(preds, model, metric, mod = NULL)
}
\arguments{
\item{preds}{output from \code{pull_sdm_preds}}

\item{model}{one of the following indicating the desired model to calculate the metric for from: gam, maxent, brt, rf, sdmtmb, or ens}

\item{metric}{either rmse or auc - determines which metric is calculated}

\item{mod}{the output from \code{build_sdm} - only used for ensemble}
}
\value{
the desired evaluation metric for the given model
}
\description{
Calculate RMSE or AUC for one of the component models or the final ensemble
}
