% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edges.R
\name{get_edges}
\alias{get_edges}
\title{Calculate leading and trailing edges of species distributions}
\usage{
get_edges(coords, weights, probs = c(0.05, 0.95))
}
\arguments{
\item{coords}{a vector of x or y coordinates}

\item{weights}{a vector of sdm values predicted at x/y coordinates.}

\item{probs}{the desired probabilities to calculate as leading and trailing edges. defaults to 5 and 95\%.}
}
\value{
a vector with length equal to the number of probabilities supplied with the x or y values corresponding to the provided probabilities
}
\description{
calculate trailing and leading edges of species distributions using cumulative density function
}
