% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kde.R
\name{get_kde}
\alias{get_kde}
\title{Calculate leading and trailing edges of species distributions}
\usage{
get_kde(r_single, level)
}
\arguments{
\item{r_single}{a spatRaster with one layer}

\item{level}{the probability to calculate KDE for}
}
\value{
a polygon representing the KDE at the given probability
}
\description{
calculate trailing and leading edges of species distributions using cumulative density function
}
