% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMOM6Forecast.R
\name{getMOM6Forecast}
\alias{getMOM6Forecast}
\title{Pull MOM6 Forecast Data}
\usage{
getMOM6Forecast(varDF, inPar = TRUE, jsonURL, release, init, ens)
}
\arguments{
\item{varDF}{a data.frame with the columns Long.Name and Short.Name to be used as reqVars and shortNames, respectively}

\item{inPar}{TRUE/FALSE to determine if data pulls and averaging should be conducted in parallel with dopar}

\item{jsonURL}{URL pointing to JSON table variable lists for desired MOM6 run type and domain}

\item{release}{release code. Must match one of the options in the 'cefi_release' column in provided JSON table}

\item{init}{initialization code. Must match one of the options in the 'cefi_init_date' column in provided JSON table. For forecast only}

\item{ens}{ensemble member. Must be equal to 1-10. For decadal forecasts, different ensemble members represent slightly different forcing scenarios. For forecast only.}
}
\value{
the output from \code{norm_env} - a list whose length is equal to the number of variables supplied, where each item in the list is a rasterStack of data associated with that variable, with the number of layers equal to the number of time steps available. The function also saves the output from each step - see the package website for necessary directory set up.
}
\description{
\code{getMOM6Forecast} is a wrapper function that pulls, averages, and normalizes the MOM6 forecast data using the \code{pull_forecast}, \code{avg_env}, \code{sd_env}, and \code{norm_env} functions, and saves the output from each step. This function adds log functionality to help with batch runs and running the encompassing functions in parallel. Note that unlike similar functions, there is no skip functionality since this function is meant to only be run once.
}
