% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_exposure_plots.R
\name{make_exposure_plots}
\alias{make_exposure_plots}
\title{Make Plots of Exposure Results}
\usage{
make_exposure_plots(
  species,
  type = c("variable", "total", "important", "radar"),
  forecast_release,
  forecast_init,
  hindcast_release,
  hindcast_yr_range,
  variable_df,
  coastline,
  bathymetry
)
}
\arguments{
\item{species}{names of the species to plot. Must match folder name to pull correct data and save figures correctly.}

\item{type}{at least one of the following: 'variable', 'total', 'important', 'radar'. Used to determine what to plot. Defaults to all.}

\item{forecast_release, hindcast_release}{MOM6 release codes for the (f)orecast and (h)indcasts used. Used to pull correct variable exposures}

\item{forecast_init}{forecast_initialization code corresponding to the forecast_initalization date of the desired forecast data. Used to pull correct variable exposures}

\item{hindcast_yr_range}{character string corresponding to the years in the hindcast data used. Used to pull correct data and save the data properly}

\item{variable_df}{a data.frame containing all possible environmental variables, such as from the MOM6 model. Must contain columns \code{Long.Name} and \code{Short.Name}, containing the full names and abbreviated names of the variables. Abbreviated names should correspond to those in the weights vector produced by \code{combineWeights}}

\item{coastline}{shapefile used to plot land in model prediction plots}

\item{bathymetry}{spatRaster file of bathymetry data; used to plot bathymetry in stock boundary plots}
}
\value{
Function does not return anything. Figures are saved to species-specific \code{figures} folder.
}
\description{
Produces exposure plots. Requires directory to be set up per directions in the package documentation/manual.
}
