% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pa_df_wrapper.R
\name{make_pa_df_wrapper}
\alias{make_pa_df_wrapper}
\title{Make Presence/Absence Data Frame for Modeling}
\usage{
make_pa_df_wrapper(name, skip, m_min, m_max, y_min, y_max)
}
\arguments{
\item{name}{Species name to add to log files and save data to correct directory (see vignette for recommended directory set up)}

\item{skip}{TRUE/FALSE indicating whether to skip creating the raster file if file already exists}

\item{m_min, m_max, y_min, y_max}{(m)onth and (y)ear ranges to help create names for combined rasterBrick since they aren't always retained and these are necessary for matching fisheries and environmental data}
}
\value{
does not return anything to the environment; it only saves the data sets generated by each step and writes to the log file
}
\description{
This is a wrapper function for \code{merge_spp_env}, \code{match_guilds}, \code{remove_corr}, and \code{set_binary_pa} that produces logs and has skip functionality to assist in running multiple species in parallel.
}
