% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{make_predictions}
\alias{make_predictions}
\title{Predict Component or Ensemble SDM}
\usage{
make_predictions(
  mod,
  model,
  rasts,
  staticVars,
  bathyR,
  mask = T,
  bathy_max,
  se = NULL,
  month_col,
  year_col,
  xy_col,
  weights = NULL
)
}
\arguments{
\item{mod}{the output from \code{make_sdm} - only used for ensemble}

\item{model}{one of the following indicating the desired model to calculate variable importance for: gam, maxent, brt, rf, sdmtmb, or ens}

\item{rasts}{list of rasterStacks corresponding to the environmental covariates used to build the models. The number of layers in each rasterStack should be the same and correspond to the length of the timeseries for the models to be predicted on}

\item{staticVars}{list of rasters containing the static variables used in model. Should be the same object used in \code{merge_spp_env}.}

\item{bathyR, bathy_max}{a raster of bathymetry with the same extent and resolution as \code{rasts} and the maximum depth you want included. For example, if you want to mask off waters deeper than 1000 m, \code{bathy_max} would be set to 1000. The value should be positive regardless of the sign of your bathymetry data.}

\item{mask}{TRUE/FALSE indicating whether to mask off certain depths (i.e. waters deeper than 1000 m)}

\item{se}{data frame containing species presence/absence data and desired environmental covariate data.}

\item{month_col, year_col}{column names for month and year columns respectively}

\item{xy_col}{a vector with a length of 2 indicating the longitude and latitude column names}

\item{weights}{a vector of model weights - used for building the ensemble model}
}
\value{
returns a rasterStack of predicted habitat suitability. The number of layers will be equal to the number of layers in \code{rasts}
}
\description{
Make model predictions for either a component model (GAM/MAXENT/RF/BRT/SDMTMB) or the Ensemble model
}
