% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sdm_plots.R
\name{make_sdm_plots}
\alias{make_sdm_plots}
\title{Make Plots for Species Distribution Models}
\usage{
make_sdm_plots(
  species,
  type = c("ensemble", "weights", "importance", "residuals"),
  release,
  spatial_temporal,
  mask_bathy,
  rm_corr,
  training_years,
  source = "hindcast",
  init = NA,
  var_names,
  component_models,
  coastline,
  model_metrics
)
}
\arguments{
\item{species}{name of the species to plot. Must match folder name to pull correct data and save figures correctly.}

\item{type}{at least one of the following: 'ensemble', 'weights', 'importance', 'residuals'. Used to determine what to plot.}

\item{release}{release code for MOM6 data. Helps pull correct model projections associated with the MOM6 data with the same name}

\item{spatial_temporal}{TRUE/FALSE to determine method for normalizing. Helps pull correct model projections associated with the MOM6 data with the same name}

\item{mask_bathy}{TRUE/FALSE indicating whether or not bathymetry data was used as a mask for raw data before normalization. Helps pull correct model projections associated with the MOM6 data with the same name}

\item{rm_corr}{TRUE/FALSE indicating whether or not correlated environmental covariates were removed. Helps to pull correct model projections}

\item{training_years}{vector with length equal to 2, indicating the maximum and minimum years that identify the desired training datasets}

\item{source}{'hindcast' or 'forecast' to determine which MOM6 data to pull}

\item{init}{initalization code for forecast model, defaults to NA since it is not needed for hindcast}

\item{var_names}{character vector containing possible variable names. matched against names in data.frame for variable importance extraction. should include both dynamic (environmental) and static (geographical) names.}

\item{component_models}{character vector containing the component models to use for weight and importance plotting.}

\item{coastline}{shapefile used to plot land in model prediction plots}

\item{model_metrics}{a data.table containing the model metrics for each species. Generated by \code{make_evaluation_csv}.}
}
\value{
Function does not return anything. Figures are saved to species-specific \code{figures} folder.
}
\description{
Produces plots necessary for SDM reports. Requires directory to be set up per directions in the package documentation/manual.
}
