% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sdm_predictions.R
\name{make_sdm_predictions}
\alias{make_sdm_predictions}
\title{Predict Component or Ensemble SDM}
\usage{
make_sdm_predictions(
  mod,
  model,
  rasts,
  static_variables,
  se = NULL,
  month_col,
  pa_col,
  year_col,
  xy_col,
  weights = NULL
)
}
\arguments{
\item{mod}{the output from \code{make_sdm} - only used for ensemble}

\item{model}{one of the following indicating the desired model to calculate variable importance for: gam, maxent, brt, rf, sdmtmb, or ens}

\item{rasts}{for component models (GAM/MAXENT/RF/BRT/SDMTMB) list of spatRasters corresponding to the environmental covariates used to build the models. The number of layers in each spatRaster should be the same and correspond to the length of the timeseries for the models to be predicted on. For the Ensemble, a list of spatRasters containing the predicted values for each of the component models. The length of the list should be equal to the length of \code{weights}}

\item{static_variables}{spatRaster containing the static variables used in model.}

\item{se}{data frame containing species presence/absence data and desired environmental covariate data.}

\item{month_col, year_col}{column names for month and year columns respectively}

\item{pa_col}{column name for presence/absence column}

\item{xy_col}{a vector with a length of 2 indicating the longitude and latitude column names}

\item{weights}{a vector of model weights - used for building the ensemble model}
}
\value{
returns a rasterStack of predicted habitat suitability. The number of layers will be equal to the number of layers in \code{rasts}
}
\description{
Make model predictions for either a component model (GAM/MAXENT/RF/BRT/SDMTMB) or the Ensemble model
}
