% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sensitivity_barplots.R
\name{make_sensitivity_barplots}
\alias{make_sensitivity_barplots}
\title{Make Sensitivity Barplot Reports}
\source{
This function was originally written as two seperate functions to generate 1) a PDF report of all scores and 2) individual reports for each scorer for their assigned species. These were originally written by Megan Stachura in December 2014 for the original CVA analysis in the Northeast US. The functions were combined in 2026 by the package author.
}
\usage{
make_sensitivity_barplots(
  data,
  species,
  plots_name,
  sensitivity = TRUE,
  plot_data_quality = FALSE,
  plot_legend = TRUE,
  preliminary = TRUE,
  scorer = TRUE,
  scorer_dir
)
}
\arguments{
\item{data}{all data exported from database}

\item{species}{list of all the species in the order you want them to appear in the plots}

\item{plots_name}{desired name of plot file. file extension not needed.}

\item{sensitivity}{TRUE indicates sensitivity attribute scores should be plotted. FALSE indicates exposure factor scores should be plotted}

\item{plot_data_quality}{TRUE indicates the average data quality for each attribute should be printed on the plot, FALSE indicates this should not be printed on the plot}

\item{plot_legend}{TRUE indicates a legend of the scorers should be printed on the plots, FALSE indicates this legend should not be printed}

\item{preliminary}{TRUE indicates that the "Preliminary Results" label should be printed at the top of the plots, FALSE indicates this should not be printed on the plots}

\item{scorer}{TRUE indicates that individual PDF reports for each scorer should also be produced.}

\item{scorer_dir}{the directory to save individual scorer PDF reports to}
}
\value{
No returns. Saves pdf file of the stacked barplots, with one page of barplots per species scored by the corresponding experts. PDFs will have naming convention FIRSTNAME_LASTNAME.pdf and be saved in the designated folder name
}
\description{
Generates a set of PDFs, with each page containing stacked barplots illustrating the distribution of tallies for each sensitivity attribute for each species scored. The option to generate scorer-specific reports is available.
}
