% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stock_polygons.R
\name{make_stock_polygons}
\alias{make_stock_polygons}
\title{Make Stock Polygons}
\usage{
make_stock_polygons(
  key,
  species_col,
  stock_col,
  id_col,
  polygons,
  poly_id,
  plot = T,
  bathymetry,
  coastline
)
}
\arguments{
\item{key}{data.frame listing all of the desired species, stocks, and associated polygons in an ID column}

\item{species_col, stock_col}{character string matching columns in \code{key} containing species and stock names}

\item{id_col}{character string matching the ID column in \code{key} use to match to polygon names}

\item{polygons}{a spatVector containing all the potential polygons to match to}

\item{poly_id}{a character string matching the ID column in \code{polygons} to match polygons}

\item{plot}{TRUE/FALSE turning on/off plotting code}

\item{bathymetry}{spatRaster file of bathymetry data; used to plot bathymetry in stock boundary plots}

\item{coastline}{shapefile used to plot land in stock boundary plots}
}
\value{
saves species-specific spatVector to working directory. returns log of species and stocks found in key
}
\description{
Make polygons for each stock for a species based on provided data.frame. Saves
}
