% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDF.R
\name{makeDF}
\alias{makeDF}
\title{Make Data Frame}
\usage{
makeDF(name, skip, mMin, mMax, yMin, yMax)
}
\arguments{
\item{name}{Species name to add to log files and save data to correct directory (see vignette for recommended directory set up)}

\item{skip}{TRUE/FALSE indicating whether to skip creating the raster file if file already exists}

\item{mMin, mMax, yMin, yMax}{(m)onth and (y)ear ranges to help create names for combined rasterBrick since they aren't always retained and these are necessary for matching fisheries and environmental data}
}
\value{
does not return anything to the environment; it only saves the data sets generated by each step and writes to the log file
}
\description{
This is a wrapper function for \code{merge_spp_env}, \code{match_guilds}, \code{remove_corr}, and \code{clean_data} that produces logs and has skip functionality to assist in running multiple species in parallel.
}
