% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEns.R
\name{makeEns}
\alias{makeEns}
\title{Build and Predict Ensemble Model}
\usage{
makeEns(spp, yrMin, yrMax, buildEns = T)
}
\arguments{
\item{spp}{Species name to add to log files and save data to correct directory (see vignette for recommended directory set up)}

\item{yrMin}{start of year range to specify which predictions to use}

\item{yrMax}{end of year range to specify which predictions to use}

\item{buildEns}{TRUE/FALSE to switch on/off ensemble model building, set to FALSE to just predict ensemble}
}
\value{
returns the file path of the saved predictions. Predicted rasters, model weights, and preds from the ensemble are saved within specific directories. See the vignette for recommended directory set up.
}
\description{
A wrapper function to build and predict the ensemble models using \code{make_sdm} and \code{make_predictions}. Progress is tracked via a log file. NOTE that this function does NOT include skip functionality. This is because it is quicker than the other predictions since it is performing a weighted average of the individual model predictions.
}
