% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExposureMaps.R
\name{makeExposureMaps}
\alias{makeExposureMaps}
\title{Make Maps of Species-Specific Exposure for each variable}
\usage{
makeExposureMaps(rankExp, sdmRast)
}
\arguments{
\item{rankExp}{output from \code{rankExposure}}

\item{sdmRast}{rasterStack with monthly averages of SDM results to use as weights for weighted average}
}
\value{
a rasterStack with the number of layers equal to the number of variables supplied, containing the weighted average exposure, weighted by SDM results, for each variable.
}
\description{
Combines raw variable exposures and SDM results, using the SDM results as weights to average the exposure across time to produce maps of species-specific exposure for each variable
}
