% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExposureTimeseries.R
\name{makeExposureTimeseries}
\alias{makeExposureTimeseries}
\title{Make Timeseries of Species-Specific Exposure for each variable}
\usage{
makeExposureTimeseries(rankExp, sdmRast)
}
\arguments{
\item{rankExp}{output from \code{rankExposure}}

\item{sdmRast}{rasterStack with monthly averages of SDM results to use as weights for weighted average}
}
\value{
A matrix with the number of rows equal to the number of variables supplied, and 12 columns (1 for each month) containing the weighted average exposure, weighted by SDM results, for each variable.
}
\description{
Combines raw variable exposures and SDM results, using the SDM results as weights to average the exposure across space to produce timeseries of species-specific exposure for each variable
}
