% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMods.R
\name{makeMods}
\alias{makeMods}
\title{Make Component Models}
\usage{
makeMods(spp, model, skip)
}
\arguments{
\item{spp}{Species name to add to log files and save data to correct directory (see vignette for recommended directory set up)}

\item{model}{Component model name. One of the following: gam, maxent, rf, brt, sdmtmb}

\item{skip}{TRUE/FALSE indicating whether to skip creating the raster file if file already exists}
}
\value{
returns a summary object of the model built. Outputs from the subsequent functions called within are saved within specific directories. See the vignette for recommended directory set up.
}
\description{
This is a wrapper function for all of the functions used to build and evaluate the component models: \code{make_sdm}, \code{sdm_cv}, \code{sdm_preds}, \code{sdm_eval}, and \code{sdm_importance}. This function produces log files and has skip functionality to assist in running multiple species in parallel.
}
