% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePredDF.R
\name{makePredDF}
\alias{makePredDF}
\title{Convert Environmental Rasters to a Data.Frame}
\usage{
makePredDF(rasts, staticVars, bathyR, bathy_max, mask)
}
\arguments{
\item{rasts}{list of rasterStacks corresponding to the environmental covariates used to build the models. The number of layers in each rasterStack should be the same and correspond to the length of the timeseries for the models to be predicted on}

\item{staticVars}{list of rasters containing the static variables used in model. Should be the same object used in \code{merge_spp_env}.}

\item{bathyR, bathy_max}{a raster of bathymetry with the same extent and resolution as \code{rasts} and the maximum depth you want included. For example, if you want to mask off waters deeper than 1000 m, \code{bathy_max} would be set to 1000. The value should be positive regardless of the sign of your bathymetry data.}

\item{mask}{TRUE/FALSE indicating whether to mask off certain depths (i.e. waters deeper than 1000 m)}
}
\value{
a data.frame containing all of the environmental data in \code{rasts} for all timesteps. It will be big depending on the length of your time series. This is meant to be used to help predict sdmTMB models.
}
\description{
Make a data frame out of environmental data to be used for predictions. Helps speed up sdmTMB predictions.
}
