% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePredRasters.R
\name{makePredRasters}
\alias{makePredRasters}
\title{Convert Predicted Values in Data.Frame to Rasters}
\usage{
makePredRasters(df, staticVars)
}
\arguments{
\item{df}{the output from \code{makePredDF}}

\item{staticVars}{list of rasters containing the static variables used in model. Should be the same object used in \code{merge_spp_env}. Used in this instance to set resolution of rasters}
}
\value{
a rasterStack with the same number of layers as in \code{rasts} with probability of occurance predicted for each day with available data. Values will range from 0 to 1.
}
\description{
Make rasters from predicted values in a data.frame. Helps speed up sdmTMB predictions.
}
