% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makereport_exposuretable.R
\name{makereport_exposuretable}
\alias{makereport_exposuretable}
\title{Make Exposure Summary Table}
\usage{
makereport_exposuretable(
  species,
  presentTime,
  futureTime,
  variableDF,
  model.metrics,
  auc.col,
  model.confidence
)
}
\arguments{
\item{species}{names of the species to plot. Must match folder name to pull correct data and save table correctly.}

\item{presentTime, futureTime}{character strings indicating the present and future time series to compare. Example: '1993-2019'. Used to pull correct calculations and save the data properly}

\item{variableDF}{a data.frame containing all possible environmental variables, such as from the MOM6 model. Must contain columns \code{Long.Name} and \code{Short.Name}, containing the full names and abbreviated names of the variables. Abbreviated names should correspond to those in the weights vector produced by \code{combineWeights}}

\item{model.metrics}{a data.table containing the model metrics for each species. Generated by \code{model.metrics}.}

\item{auc.col}{name of the column containing the AUC values to be included in the table footnote}

\item{model.confidence}{a data.table containing the model confidence scores for each species.}
}
\value{
Function does not return anything. The table is saved to species- and timeframe-specific \code{figures} folder.
}
\description{
Generates the summary exposure table with final factor scores, standard deviations, and distribution of spatial and temporal for each species, as well as final exposure scores. One table is generated per species and saved in the species folder. Requires directory to be set up per directions in the package documentation/manual.
}
