% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makereports_sdms.R
\name{makereport_sdms}
\alias{makereport_sdms}
\title{Render Species Distribution Model (SDM) reports}
\usage{
makereport_sdms(
  species.list,
  yrStart,
  yrEnd,
  model.metrics,
  feeding_key,
  habitat_key,
  variable_key,
  template,
  report.path
)
}
\arguments{
\item{species.list}{a data.table containing the species names. Must contain the column \code{Common.name}}

\item{yrStart, yrEnd}{start and end of timeseries to plot, used to pick correct model predictions to include}

\item{model.metrics}{a data.table containing the model metrics for each species. Generated by \code{model.metrics}.}

\item{feeding_key, habitat_key}{file path to csv for feeding and habitat guild keys respectively. Used for matching variable names. See vignette for recommended set up of these files.}

\item{variable_key}{a data.frame containing all possible modeled variables, such as from the MOM6 model. Must contain columns \code{Long.Name} and \code{Short.Name}, containing the full names and abbreviated names of the variables. Abbreviated names should correspond to those in the variable importance plots}

\item{template}{a character string designating the .qmd file to use as a template}

\item{report.path}{character string designating where the reports should be saved}
}
\value{
Function does not return anything. Reports are saved to \code{report.path}
}
\description{
Builds SDM report using a provided template. Requires directory to be set up per directions in the package documentation/manual.
}
