% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makereport_sensitivitybarplots.R
\name{makereport_sensitivitybarplots}
\alias{makereport_sensitivitybarplots}
\title{Make Scorer Barplot Reports}
\source{
This function was originally written by Megan Stachura in December 2014 for the original CVA analysis in the Northeast US. It was tested by the package author in 2026 and is still functional.
}
\usage{
makereport_sensitivitybarplots(
  data,
  species,
  plots.file.name,
  sensitivity = TRUE,
  plotDataQuality = FALSE,
  plotLegend = TRUE,
  preliminary = TRUE
)
}
\arguments{
\item{data}{all data exported from database}

\item{species}{list of all the species in the order you want them to appear in the plots}

\item{plots.file.name}{desired name of plot file. file extension not needed.}

\item{sensitivity}{TRUE indicates sensitivity attribute scores should be plotted. FALSE indicates exposure factor scores should be plotted}

\item{plotDataQuality}{TRUE indicates the average data quality for each attribute should be printed on the plot, FALSE indicates this should not be printed on the plot}

\item{plotLegend}{TRUE indicates a legend of the scorers should be printed on the plots, FALSE indicates this legend should not be printed}

\item{preliminary}{TRUE indicates that the "Preliminary Results" label should be printed at the top of the plots, FALSE indicates this should not be printed on the plots}
}
\value{
No returns. Saves pdf file of the stacked barplots, with one page of barplots per species scored by the corresponding experts. PDFs will have naming convention FIRSTNAME_LASTNAME.pdf and be saved in the designated folder name
}
\description{
Generates a set of PDFs, with each page containing stacked barplots illustrating the distribution of tallies for each sensitivity attribute for each species scored by an expert. One report is generated per expert in the designated directory.
}
