% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makereport_sensitivitytable.R
\name{makereport_sensitivitytable}
\alias{makereport_sensitivitytable}
\title{Make Sensitivity Table}
\usage{
makereport_sensitivitytable(
  species,
  species_col,
  attribute_names_raw,
  attribute_names_clean,
  total_sens_col,
  certainty_col,
  raw.data,
  sensitivity,
  data.quality,
  table.folder
)
}
\arguments{
\item{species}{names of species to produce table for. Must match \code{Stock.Name} in \code{raw.data}}

\item{species_col}{name of column with species names to help subset data. Should be the same across \code{raw.data}, \code{sensitivity}, and \code{data.quality}}

\item{attribute_names_raw, attribute_names_clean}{vectors of raw attribute names to help subset \code{sensitivity} and \code{data.quality}, and cleaned up names to use in table}

\item{total_sens_col, certainty_col}{names of column containing total sensitivity scores and associated certainty in \code{sensitivity}}

\item{raw.data}{raw data exported from FCVA database, including the column \code{Stock.Name} to subset data to correct species}

\item{sensitivity}{TRUE indicates sensitivity attribute scores should be plotted. FALSE indicates exposure factor scores should be plotted}

\item{data.quality}{data.frame with data quality scores for each attribute and species. Generated by \code{data.quality.score}. Must include the column \code{Stock.Name} to subset data to correct species}

\item{table.folder}{file path to folder to save tables in}
}
\value{
No returns. Saves pdf file containing the sensitivity table containing the expert scores, data quality, and overall sensitivity score, with certainty.
}
\description{
Generates the sensitivity table with final attribute scores, data quality, and distribution of tallies for each species. One table is generated per species and saved in the designated directory. Requires directory to be set up per directions in the package documentation/manual.
}
