% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeVariableAverages.R
\name{makeVariableAverages}
\alias{makeVariableAverages}
\title{Make Timeseries and Maps of Species-Specific Exposure for each variable}
\usage{
makeVariableAverages(spp, ensName, sdmThreshold = 0.1, presentTime, futureTime)
}
\arguments{
\item{spp}{species name. Used to pull correct data and save outputs in species-specific folders.}

\item{ensName}{name of ensemble species distribution model file to pull}

\item{sdmThreshold}{value between 0 and 1. Will remove values lower than this threshold from average ensemble model results to help reduce weird aliasing that can occur in workflow. Defaults to 0.1.}

\item{presentTime, futureTime}{character strings indicating the present and future time series to compare. Example: '1993-2019'. Used to pull correct ranked exposure values and save the data properly}
}
\value{
Nothing is returned. The outputs from \code{makeExposureMaps} and \code{MakeExposureTimeseries} are saved in the appropriate folders
}
\description{
A wrapper function for \code{makeExposureMaps} and \code{MakeExposureTimeseries} that handles object loading, and generating mean SDMs from timeseries
}
