% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_fisheries_environment_data.R
\name{match_fisheries_environment_data}
\alias{match_fisheries_environment_data}
\title{Match Species Presence-Absence Data Frame to Environmental Data}
\usage{
match_fisheries_environment_data(
  species_df,
  model_data,
  add_static = TRUE,
  static_variables
)
}
\arguments{
\item{species_df}{A data.frame containing species presence-absence data.}

\item{model_data}{A list containing spatRasters of each environmental variable to be matched, or the output of \code{normalize_model_data} or \code{get_model_data_wrapper}. Each member of the list should have a name which will correspond to the environmental variable's column name in the resulting dataframe.}

\item{add_static}{TRUE/FALSE to add static variables such as bathymetry. If TRUE, static_variables needs to be supplied. Note that if the user is supplying normalized dynamic environmental variable that static variables should also be normalized with the same methods.}

\item{static_variables}{a spatRaster containing static variables such as bathymetry. Names of each layer will become the column names.}
}
\value{
\code{match_fisheries_environment_data} returns a data frame with the following columns:
\itemize{
\item \code{x, y, month, year} indicating longitude, latitude, month, and year
\item \code{pa} - the column indicating presence, absence, and effort from the fisheries raster
\item A number of columns equal to the length of \code{model_data} with the same names containing the corresponding environmental data at each location
\item A number of columns equal to the length of \code{static_variables}, if \code{add_static = T}, with column names matching the names of the rasters in the list
}
}
\description{
Match species data frame and environmental rasters to build data frame for modeling
}
