% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pa_model_rasters.R
\name{match_pa_model_rasters}
\alias{match_pa_model_rasters}
\title{Match Species and Model Rasters and Create a Data Frame}
\usage{
match_pa_model_rasters(
  pa_rasters,
  model_data,
  add_static = TRUE,
  static_variables
)
}
\arguments{
\item{pa_rasters}{rasterStack representing fisheries presence, absence, and effort with a layer for each timestep. Layers should be named by timestamp, for example, month.year}

\item{model_data}{A list containing rasterStacks of each environmental variable to be matched, or the output of \code{norm_env}. Each member of the list should have a name which will correspond to the environmental variable's column name in the resulting dataframe. Each rasterStack should have the same number of layers as \code{pa_rasters}, and each layer should have a name that can be matched to the names in \code{pa_rasters}}

\item{add_static}{TRUE/FALSE to add static variables such as bathymetry. If TRUE, staticData needs to be supplied.}

\item{static_variables}{list of rasters containing static variables such as bathymetry. Names of the listed rasters will become the column names.}
}
\value{
\code{merge_pa_model_rasters} returns a data frame with the following columns:
\itemize{
\item \code{x, y, month, year} indicating longitude, latitude, month, and year
\item \code{value} - the column indicating presence, absence, and effort from the fisheries raster
\item A number of columns equal to the length of \code{model_data} with the same names containing the corresponding environmental data at each location
\item A number of columns equal to the length of \code{static_variables}, if \code{add_static = T}, with column names matching the names of the rasters in the list
}
}
\description{
Match species and environmental rasters to build data frame for modeling
}
