% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fisheries_dfs.R
\name{merge_fisheries_dfs}
\alias{merge_fisheries_dfs}
\title{Merge Fisheries Presence/Absence/Effort Data Frames}
\usage{
merge_fisheries_dfs(df_dir)
}
\arguments{
\item{df_dir}{full path to directory containing the data.frames to merge}
}
\value{
returns a single data.frame combining presence/absence data.frames from all sources. Data are aggregated within grid cells for each month and year such that each grid cell reflects if a species was present or absent within that grid cell across data sources
}
\description{
Combines data.frames from \code{build_fisheries_df} across multiple sources
}
