% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fisheries_rasters.R
\name{merge_fisheries_rasters}
\alias{merge_fisheries_rasters}
\title{Merge Fisheries Presence/Absence/Effort Rasters}
\usage{
merge_fisheries_rasters(raster_list)
}
\arguments{
\item{raster_list}{list of outputs from \code{create_rast} to combine}
}
\value{
returns a single rasterBrick with the same spatial and temporal resolution as provided rasters. Maximum value should be 2 to ensure that presences are included.
}
\description{
Combines rasters from \code{create_rast} across multiple sources
}
