% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rast.R
\name{merge_rasts}
\alias{merge_rasts}
\title{Merge Fisheries Presence/Absence/Effort Rasters}
\usage{
merge_rasts(rastList)
}
\arguments{
\item{rastList}{list of outputs from \code{create_rast} to combine}
}
\value{
returns a single rasterBrick with the same spatial and temporal resolution as provided rasters. Maximum value should be 2 to ensure that presences are included.
}
\description{
Combines rasters from \code{create_rast} across multiple sources
}
