% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_spp_env.R
\name{merge_spp_env}
\alias{merge_spp_env}
\title{Merge Species and Environmental Rasters into a Data Frame}
\usage{
merge_spp_env(rastStack, envData, addStatic = TRUE, staticVars)
}
\arguments{
\item{rastStack}{rasterStack representing fisheries presence, absence, and effort with a layer for each timestep. Layers should be named by timestamp, for example, month.year}

\item{envData}{output from one of the MOM6 data functions, preferably \code{norm_env}. A list containing rasterStacks of each environmental variable to be matched. Each member of the list should have a name which will correspond to the environmental variable's column name in the resulting dataframe. Each rasterStack should have the same number of layers as \code{rastStack}, and each layer should have a name that can be matched to the names in \code{rastStack}}

\item{addStatic}{TRUE/FALSE to add static variables such as bathymetry. If TRUE, staticData needs to be supplied.}

\item{staticVars}{list of rasters containing static variables such as bathymetry. Names of the listed rasters will become the column names.}
}
\value{
\code{merge_spp_env} returns a data frame with the following columns:
\itemize{
\item \code{x, y, month, year} indicating longitude, latitude, month, and year
\item \code{value} - the column indicating presence, absence, and effort from the fisheries raster
\item A number of columns equal to the length of \code{envData} with the same names containing the corresponding environmental data at each location
\item A number of columns equal to the length of \code{staticData}, if \code{addStatic = T}, with column names matching the names of the rasters in the list
}
}
\description{
Match species and environmental rasters to build data frame for modeling
}
