% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_env.R
\name{norm_env}
\alias{norm_env}
\title{Calculate Standard Deviation on MOM6 Data}
\usage{
norm_env(rawList, avgList, sdList, shortNames)
}
\arguments{
\item{rawList}{List of output rasters from \code{pull_hind} or \code{pull_forecast}}

\item{avgList, sdList}{List of output rasters from \code{avg_env} and \code{sd_env}, respectively}

\item{shortNames}{vector of simplified variable names to help name resulting raster files.}
}
\value{
a list whose length is equal to the number of variables supplied, where each item in the list is a rasterStack of data associated with that variable
}
\description{
Calculate monthly standard deviation from raw MOM6 data across the entire provided timeseries
}
