% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_model_data.R
\name{normalize_model_data}
\alias{normalize_model_data}
\title{Normalize Model Data}
\usage{
normalize_model_data(raw, avg, sd, spatial_temporal)
}
\arguments{
\item{raw}{Output spatRaster from \code{pull_mom6_hindcast} or \code{pull_mom6_forecast}. Alternatively, a spatRaster of data to be normalized}

\item{avg, sd}{Output spatRaster from \code{avg_model_data} and \code{sd_model_data}, respectively. Alternatively, averages and standard deviations of desired gridded datasets.}

\item{spatial_temporal}{TRUE/FALSE to determine how normalization should occur. If TRUE, calculations should be spatially and temporally explicit. \code{avg} and \code{sd} should be spatRaster objects with averages and standard deviations of data in space and time. If FALSE, the calculation occurs across space and time; \code{avg} and \code{sd} should be single values representing the overall average and standard deviation for the given variable.}
}
\value{
a spatRaster of normalized environmental data
}
\description{
Normalize model data across the entire provided timeseries using a z-score
}
