% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_variable_importance.R
\name{normalize_variable_importance}
\alias{normalize_variable_importance}
\title{Normalize Dynamic Variable Weights}
\usage{
normalize_variable_importance(vars, ens_weights, imp_list)
}
\arguments{
\item{vars}{character vector of variables to generate normalized variable importance for}

\item{ens_weights}{vector of component model weights in the ensemble}

\item{imp_list}{a named list of importance outputs from different models. names should match model names function will handle differences between them.}
}
\value{
A vector representing the weighted average of normalized variable importance, representing variable importance in the final ensemble SDM.
}
\description{
Normalizes the component model variable importance, as well as the model weights in the final ensemble, to create a weighted average vector of variable importance for the ensemble model. Options allow for static variables, such as time and space, or non-dynamic environmental variables such as bathymetry, to be excluded from the normalization and weighted average.
}
