% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SDMS.R
\name{plot_SDMS}
\alias{plot_SDMS}
\title{Make Plots for Species Distribution Models}
\usage{
plot_SDMS(
  species,
  type = c("ensemble", "weights", "importance", "residuals"),
  yrStart,
  yrEnd,
  coastline,
  bathy,
  model.metrics
)
}
\arguments{
\item{species}{name of the species to plot. Must match folder name to pull correct data and save figures correctly.}

\item{type}{at least one of the following: 'ensemble', 'weights', 'importance', 'residuals'. Used to determine what to plot.}

\item{yrStart, yrEnd}{start and end of timeseries to plot, used to pick correct model predictions}

\item{coastline}{shapefile used to plot land in model prediction plots}

\item{bathy}{raster file of bathymetry data, used to help mask off deeper predictions}

\item{model.metrics}{a data.table containing the model metrics for each species. Generated by \code{make_performanceCSV}.}
}
\value{
Function does not return anything. Figures are saved to species-specific \code{figures} folder.
}
\description{
Produces plots necessary for SDM reports. Requires directory to be set up per directions in the package documentation/manual.
}
