% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sdms_wrapper.R
\name{predict_sdms_wrapper}
\alias{predict_sdms_wrapper}
\title{Predict Models}
\usage{
predict_sdms_wrapper(spp, model, yr_min, yr_max, skip)
}
\arguments{
\item{spp}{Species name to add to log files and save data to correct directory (see vignette for recommended directory set up)}

\item{model}{Component model name. One of the following: gam, maxent, rf, brt, sdmtmb}

\item{yr_min, yr_max}{start/end of year range of predictions; adds range to resulting file name and helps pull appropriate timeseries for plotting and exposure analysis}

\item{skip}{TRUE/FALSE indicating whether to skip creating the raster file if file already exists}
}
\value{
returns the file path of the saved predictions. Outputs are saved within specific directories. See the vignette for recommended directory set up.
}
\description{
The wrapper function for \code{make_sdm_predictions}. This function produces log files and has skip functionality to assist in running multiple species in parallel.
}
