% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMods.R
\name{predictMods}
\alias{predictMods}
\title{Predict Models}
\usage{
predictMods(spp, model, skip)
}
\arguments{
\item{spp}{Species name to add to log files and save data to correct directory (see vignette for recommended directory set up)}

\item{model}{Component model name. One of the following: gam, maxent, rf, brt, sdmtmb}

\item{skip}{TRUE/FALSE indicating whether to skip creating the raster file if file already exists}
}
\value{
returns the file path of the saved predictions. Outputs are saved within specific directories. See the vignette for recommended directory set up.
}
\description{
The wrapper function for \code{make_predictions}. This function produces log files and has skip functionality to assist in running multiple species in parallel.
}
