% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_time_step_df.R
\name{prep_time_step_df}
\alias{prep_time_step_df}
\title{Build data.frame for single timestep}
\usage{
prep_time_step_df(x, rasts, static_variables)
}
\arguments{
\item{x}{a numeric corresponding to the layer to pull. Generated within \code{make_sdm_predictions}.}

\item{rasts}{list of environmental spatRasters to use in predictions. Each spatRaster should have the name number of layers, corresponding to different timestamps}

\item{static_variables}{spatRaster containing the static variables used in model.}
}
\value{
a data.frame containing all of the data in \code{rasts} and \code{static_variables} for a single timestep.
}
\description{
Helper function used in \code{make_sdm_predictions} to combine all environmental rasters for a single timestep and necessary static variables to predict model into a single data.frame
}
