% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_time_step_stack.R
\name{prep_time_step_stack}
\alias{prep_time_step_stack}
\title{Build spatRaster for single timestep}
\usage{
prep_time_step_stack(
  x,
  rasts,
  template_r,
  rlon,
  rlat,
  static_variables,
  month_col,
  year_col,
  xy_col
)
}
\arguments{
\item{x}{a numeric corresponding to the layer to pull. Generated within \code{make_sdm_predictions}.}

\item{rasts}{list of environmental spatRasters to use in predictions. Each spatRaster should have the name number of layers, corresponding to different timestamps}

\item{template_r}{a single-layer spatRaster to use as a template for other raster. Used to make sure \code{static_variables} have the correct resolution and projection, as well as make month/year rasters. Generated within \code{make_sdm_predictions}.}

\item{rlon, rlat}{rasters of longitude and latitude respectively. Generated within \code{make_sdm_predictions}.}

\item{static_variables}{spatRaster containing the static variables used in model.}

\item{month_col, year_col}{column names for month and year columns respectively.}

\item{xy_col}{a vector with a length of 2 indicating the longitude and latitude column names}
}
\value{
a spatRaster with a number of layers equal to the number of layers in \code{rasts} + number of layesr in \code{static_variables} + 4 (lon, lat, month, year)
}
\description{
Helper function used in \code{make_sdm_predictions} to combine all environmental rasters for a single timestep and necessary static variables to predict model
}
