% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_r2_brt.R
\name{pseudo_r2_brt}
\alias{pseudo_r2_brt}
\title{calculate Pseudo-R2 for BRT}
\source{
https://github.com/elhazen/EcoCast-SciAdv
}
\usage{
pseudo_r2_brt(x)
}
\arguments{
\item{x}{is fitted model of class xxx}
}
\value{
value representing Pseudo-R2
}
\description{
Used as part of \code{eval_brt}. From Camrin Brawn (WHOI): https://zenodo.org/records/7971532. Included with CVA2.0 package to ensure functionality.
}
