% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoR2.brt.R
\name{pseudoR2.brt}
\alias{pseudoR2.brt}
\title{calculate Pseudo-R2 for BRT}
\source{
https://github.com/elhazen/EcoCast-SciAdv
}
\usage{
pseudoR2.brt(x)
}
\arguments{
\item{x}{is fitted model of class xxx}
}
\value{
value representing Pseudo-R2
}
\description{
Used as part of \code{eval_brt}. From Camrin Brawn (WHOI): https://zenodo.org/records/7971532. Included with CVA2.0 package to ensure functionality.
}
