% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_hind.R
\name{pull_hind}
\alias{pull_hind}
\title{Pull MOM6 Hindcast Data}
\usage{
pull_hind(
  varURL,
  reqVars,
  shortNames,
  gt = "regrid",
  of = "monthly",
  bounds = c(-78, -65, 35, 45),
  static,
  release
)
}
\arguments{
\item{varURL}{URL pointing to JSON table variable lists for desired MOM6 hindcast and domain}

\item{reqVars}{vector of variable names to pull. Must match names in the 'cefi_long_name' column provided JSON table}

\item{shortNames}{vector of simplified variable names to help name resulting raster files. Must be the same length as reqVars.}

\item{gt}{desired grid type. Must match one of the options in the 'cefi_grid_type' column in provided JSON table}

\item{of}{desired output frequency. Must match one of the options in the 'cefi_output_frequency' column in provided JSON table}

\item{bounds}{xmin, xmax, ymin, ymax of desired output raster}

\item{static}{URL to static grid for MOM6}

\item{release}{release code. Must match one of the options in the 'cefi_release' column in provided JSON table}
}
\value{
a list whose length is equal to the number of variables supplied, where each item in the list is a rasterStack of data associated with that variable
}
\description{
Pull hindcast data from the MOM6 model output based on provided URL from the CEFI portal
}
