% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_mom6_forecast.R
\name{pull_mom6_forecast}
\alias{pull_mom6_forecast}
\title{Pull MOM6 Decadal Forecast Data}
\usage{
pull_mom6_forecast(
  var_url,
  req_var,
  gt = "regrid",
  of = "monthly",
  bounds = c(-78, -65, 35, 45),
  release,
  init
)
}
\arguments{
\item{var_url}{URL pointing to JSON table variable lists for desired MOM6 forecast and domain}

\item{req_var}{variable name to pull. Must match names in the 'cefi_long_name' column provided JSON table}

\item{gt}{desired grid type. Must match one of the options in the 'cefi_grid_type' column in provided JSON table}

\item{of}{desired output frequency. Must match one of the options in the 'cefi_output_frequency' column in provided JSON table}

\item{bounds}{xmin, xmax, ymin, ymax of desired output raster}

\item{release}{release code. Must match one of the options in the 'cefi_release' column in provided JSON table}

\item{init}{initialization code. Must match one of the options in the 'cefi_init_date' column in provided JSON table. For forecast only}
}
\value{
a spatRaster of data associated with the requested variable, averaged across the 10 ensemble members
}
\description{
Pull decadal forecast data from the MOM6 model output based on provided URL from the CEFI portal. Decadal forecasts currently have 10 ensemble members, which have slight variations in the forcing mechanisms and are not forced by climate projections. As such ensemble members are currently averaged.
}
