% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_sdm_preds.R
\name{pull_sdm_preds}
\alias{pull_sdm_preds}
\title{Pull Predicted Values from SDMs CV}
\usage{
pull_sdm_preds(cv, model)
}
\arguments{
\item{cv}{output from \code{cross_validate_sdm}}

\item{model}{one of the following indicating the desired model to extract predicted values from: gam, maxent, brt, rf, or sdmtmb}
}
\value{
a data.frame containing the prediction outputs from the cross-validation necessary to calculate evaluation metric
}
\description{
Extract the predictions from the cross-validation of one of the five component models
}
