% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_exposure.R
\name{rank_exposure}
\alias{rank_exposure}
\title{Rank Raw Environmental Exposure}
\usage{
rank_exposure(exposure, flip = F)
}
\arguments{
\item{exposure}{output from \code{calculate_raw_exposure}}

\item{flip}{TRUE/FALSE option to multiply exposure by -1 to ensure that positive exposure values represent negative variable change (increasing temperatures, decreasing oxygen concentrations, etc).}
}
\value{
A spatRaster with each layer containing ranked values between 1 - 4.
}
\description{
Converts raw exposure from \code{calculate_raw_exposure} to a rank between 1-4 based on thresholds used in CVA1.0. This function also flips variables as needed such that positive exposure is always negative (i.e. both increasing temperature and decreasing oxygen concentrations result in positive exposure).
}
