% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankExposure.R
\name{rankExposure}
\alias{rankExposure}
\title{Ranks Raw Exposure}
\usage{
rankExposure(
  exposure,
  flip = T,
  noflipList = c("bottomT", "surfaceT", "bottomArg", "MLD")
)
}
\arguments{
\item{exposure}{output from \code{calcExposure}}

\item{flip}{TRUE/FALSE option to multiply exposure by -1 to ensure that positive exposure values represent negative variable change (increasing temperatures, decreasing oxygen concentrations, etc).}

\item{noflipList}{character vector naming which vectors should NOT be flipped - counterintuitive, yes, but this list was actually shorter than the variables that needed to be flipped for NECVA2.0. Names should match the names in \code{exposure}}
}
\value{
A list of rasterStacks with each layer containing ranked values between 1 - 4. The length of the list is equal to the length of the lists supplied as \code{presentRasts} and \code{futureRasts}
}
\description{
Converts raw exposure from \code{calcExposure} to a rank between 1-4 based on thresholds used in CVA1.0. This function also flips variables as needed such that positive exposure is always negative (i.e. both increasing temperature and decreasing oxygen concentrations result in positive exposure).
}
