% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_corr.R
\name{remove_corr}
\alias{remove_corr}
\title{Remove Correlated Environmental Covariates from Species/Environmental Data Frame}
\usage{
remove_corr(df, var_names)
}
\arguments{
\item{df}{data.frame of fisheries and environmental data}

\item{var_names}{a vector containing the names of the environmental variables to check for correlation}
}
\value{
a data frame with correlated covariates removed
}
\description{
This function removed correlated covariates to prepare data for modeling
}
