% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_auc_brt.R
\name{save_auc_brt}
\alias{save_auc_brt}
\title{Calculate Area under the Curve (AUC) for BRT models}
\source{
From Camrin Brawn (WHOI): https://zenodo.org/records/7971532.
}
\usage{
save_auc_brt(truth, predicted, plot_roc = FALSE, ...)
}
\arguments{
\item{truth}{observations}

\item{predicted}{predicted values}

\item{plot_roc}{TRUE/FALSE to turn on/off plotting}

\item{...}{arguments passed to \code{ROCR::plot}}
}
\value{
an objected of class \code{performance}. Same as \code{ROCR::performance}.
}
\description{
Calculate AUC for BRT k-folds. Used within \code{eval_brt}. Included with CVA2.0 package to ensure functionality.
}
