% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rasters_wrapper.R
\name{save_rasters_wrapper}
\alias{save_rasters_wrapper}
\title{Save and Build Fisheries Rasters}
\usage{
save_rasters_wrapper(csv_name, is_obs, grid, spp, spp_names, skip)
}
\arguments{
\item{csv_name}{character string indicating which csv to use to create raster without extension (i.e. 'example' NOT 'example.csv')}

\item{is_obs}{TRUE/FALSE indicating whether or not the data is observer or similar fisheries-dependent data. Will force check to determine if species is observed at least 30 times throughout timeseries before creating raster}

\item{grid}{static link to a ncdcf object with the variables lon, lat, time - can be link to remote data - must be able to be read with nc_open}

\item{spp}{Species name to add to log files and save data to correct directory (see vignette for recommended directory set up)}

\item{spp_names}{a vector containing all possible names for the target species. Must have a length >= 1}

\item{skip}{TRUE/FALSE indicating whether to skip creating the raster file if file already exists}
}
\value{
\code{saveRast} returns the range of the rasterBrick returned by \code{build_fisheries_raster}. This should be equal to 0 2 if species is present in dataset, and 0 1 if species is not caught in dataset. This function will also save the resulting rasterBrick as a netcdf file in the species' input_rasters folder
}
\description{
A wrapper function for the \code{build_fisheries_raster} function. This function adds log and skip functionality to help with batch runs and running \code{build_fisheries_raster} in parallel.
}
