% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTSS.R
\name{saveTSS}
\alias{saveTSS}
\title{Calculate True Skill Statistic (TSS) for BRT models}
\source{
From Camrin Brawn (WHOI): https://zenodo.org/records/7971532.
}
\usage{
saveTSS(truth, predicted)
}
\arguments{
\item{truth}{observations}

\item{predicted}{predicted values}
}
\value{
a numeric value
}
\description{
Calculate TSS for BRT k-folds. Used within \code{eval_brt}. Included with CVA2.0 package to ensure functionality.
}
