% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_cv.R
\name{sdm_cv}
\alias{sdm_cv}
\title{Component Model Cross-Validation}
\usage{
sdm_cv(mod, se, pa_col, xy_col, month_col, year_col, model)
}
\arguments{
\item{mod}{the output from \code{make_sdm}}

\item{se}{data frame containing species presence/absence data and desired environmental covariate data.}

\item{pa_col}{column name for presence/absence column}

\item{xy_col}{a vector with a length of 2 indicating the longitude and latitude column names}

\item{month_col, year_col}{column names for month and year columns respectively}

\item{model}{one of the following indicating the desired model to perform cross validation on: gam, maxent, brt, rf, or sdmtmb}
}
\value{
the cross-validation results from the desired model. Objects will differ slightly depending on model type.
}
\description{
Perform cross-validation on one of the five component model types. k is 5 for all models.
}
