% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_eval.R
\name{sdm_eval}
\alias{sdm_eval}
\title{Calculate Performance Metric}
\usage{
sdm_eval(preds, model, metric, mod = NULL)
}
\arguments{
\item{preds}{output from \code{sdm_preds}}

\item{model}{one of the following indicating the desired model to calculate the metric for from: gam, maxent, brt, rf, sdmtmb, or ens}

\item{metric}{either rmse or auc - determines which metric is calculated}

\item{mod}{the output from \code{make_sdm} - only used for ensemble}
}
\value{
the desired evaluation metric for the given model
}
\description{
Calculate RMSE or AUC for one of the component models or the final ensemble
}
