% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_preds.R
\name{sdm_preds}
\alias{sdm_preds}
\title{Extract Predicted Values from Component Model CV}
\usage{
sdm_preds(cv, model)
}
\arguments{
\item{cv}{output from \code{sdm_cv}}

\item{model}{one of the following indicating the desired model to extract predicted values from: gam, maxent, brt, rf, or sdmtmb}
}
\value{
a data.frame containing the prediction outputs from the cross-validation necessary to calculate evaluation metric
}
\description{
Extract the predictions from the cross-validation of one of the five component models
}
