% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_binary_pa.R
\name{set_binary_pa}
\alias{set_binary_pa}
\title{Convert Species/Environmental Data Frame to Presence/Absence}
\usage{
set_binary_pa(se, pa_col)
}
\arguments{
\item{se}{data.frame of fisheries and environmental data}

\item{pa_col}{column name for presence/absence column}
}
\value{
a data frame where presence/absence has been set to 0 for absent and 1 for present
}
\description{
This function converts the presence/absence/effort (0-2) data to presence/absence data (0-1)
}
