% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testEns.R
\name{testEns}
\alias{testEns}
\title{Test the Ensemble Model with Observations}
\usage{
testEns(spp, sppNames, sources, yrMin, yrMax)
}
\arguments{
\item{spp}{species name. Used to help pull correct ensemble model.}

\item{sppNames}{a vector containing all possible names for the target species. Must have a length >= 1}

\item{sources}{a vector containing all of the source csvs containing the observation data to use in the test. These are loaded in directly by \code{read.csv} so they should either be full file paths or in the current working directory.}

\item{yrMin}{start of year range to specify which predictions to use}

\item{yrMax}{end of year range to specify which predictions to use}
}
\value{
returns the AUC value for the given model on the new data
}
\description{
Calculate Area under the Curve (AUC) for the ensemble species distribution model on an external set of data (not used for building the model). This function builds a dataset of presence/absence with the new data across a variety of sources, extracts the predicted values from the ensemble, and calculates the resulting AUC.
}
