% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_exposures_wrapper.R
\name{variable_exposures_wrapper}
\alias{variable_exposures_wrapper}
\title{Make Timeseries and Maps of Species-Specific Exposure for each variable}
\usage{
variable_exposures_wrapper(
  spp,
  ens_name,
  sdm_threshold = 0.1,
  present_time,
  future_time
)
}
\arguments{
\item{spp}{species name. Used to pull correct data and save outputs in species-specific folders.}

\item{ens_name}{name of ensemble species distribution model file to pull}

\item{sdm_threshold}{value between 0 and 1. Will remove values lower than this threshold from average ensemble model results to help reduce weird aliasing that can occur in workflow. Defaults to 0.1.}

\item{present_time, future_time}{character strings indicating the present and future time series to compare. Example: '1993-2019'. Used to pull correct ranked exposure values and save the data properly}
}
\value{
Nothing is returned. The outputs from \code{make_variable_exposure(type = 'map')} and \code{make_variable_exposure(type = 'timeseries')} are saved in the appropriate folders
}
\description{
A wrapper function for \code{make_variable_exposure} that handles object loading, and generating mean SDMs from timeseries, and produces both maps and timeseries
}
