% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_model_data.R
\name{avg_model_data}
\alias{avg_model_data}
\title{Average Model Data}
\usage{
avg_model_data(raw_list)
}
\arguments{
\item{raw_list}{List of output rasters from \code{pull_hind} or \code{pull_forecast}}
}
\value{
a list whose length is equal to the number of variables supplied, where each item in the list is a rasterStack of data associated with that variable
}
\description{
Calculate monthly average from raw model data across the entire provided timeseries. This is built specificially for MOM6 output, but would work on any rasterStack of gridded data.
}
