% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_evaluation_csv.R
\name{make_evaluation_csv}
\alias{make_evaluation_csv}
\title{Create a spreadsheet containing all SDM performance metrics}
\usage{
make_evaluation_csv(spp_list, test_ens = T, yr_min, yr_max)
}
\arguments{
\item{spp_list}{the data.frame containing species names and alternative names for \code{test_ens}. Must contain the column \code{Name}, which matches the names of the species folders to help pull correct data.}

\item{test_ens}{a TRUE/FALSE indicating whether or not to test the ensemble model on an external dataset}

\item{yr_min}{start of year range to specify which predictions to use. Only used if \code{test_ens == TRUE}}

\item{yr_max}{end of year range to specify which predictions to use. Only used if \code{test_ens == TRUE}}
}
\value{
nothing is returned. The resulting CSV file called 'species_evaluation_metrics.csv' is saved in the working directory.
}
\description{
Loads existing performance metrics for component models, and calculates Area under the Curve (AUC) for the ensemble model 2 ways. This function also has the option to test the model on an external dataset using \code{test_ens}. The final product is a saved csv file containing all of the performance metrics.
}
\details{
The resulting CSV file contains the following columns:
\describe{
\item{Common.Name}{name of species}
\item{Managing.Body}{name of group responsible for species management, if exists}
\item{Feeding.Guild}{assigned feeding guild}
\item{Habitat.Guild}{assigned habitat guild}
\item{N.PRESENCE, N.ABSENCE}{number of presences/absences in the data set used to build the model}
\item{BRT, GAM, MAXENT, RF, SDMTMB}{AUC for each component model}
\item{BRT.WT, GAM.WT, MAXENT.WT, RF.WT, SDMTMB.WT}{weights for each component model in the final ensemble}
\item{ENS.AUC}{Ensemble model AUC based on weighted sum of CV predictions from each component model}
\item{WAVG.ENS.AUC}{Ensemble model AUC calculated as a weighted average of AUCs, using the corresponding weights}
\item{AUC.yr_min.yr_max}{Ensemble model AUC calculated on the external data, if \code{test_ens == TRUE}. Column name will reflect the timeseries used}
}
}
