% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sdm_reports.R
\name{make_sdm_reports}
\alias{make_sdm_reports}
\title{Render Species Distribution Model (SDM) reports}
\usage{
make_sdm_reports(
  species_list,
  yr_min,
  yr_max,
  model_metrics,
  feeding_key,
  habitat_key,
  variable_key,
  template,
  report_path
)
}
\arguments{
\item{species_list}{a data.table containing the species names. Must contain the column \code{Common.name}}

\item{yr_min, yr_max}{start and end of timeseries to plot, used to pick correct model predictions to include}

\item{model_metrics}{a data.table containing the model metrics for each species. Generated by \code{model_metrics}.}

\item{feeding_key, habitat_key}{file path to csv for feeding and habitat guild keys respectively. Used for matching variable names. See vignette for recommended set up of these files.}

\item{variable_key}{a data.frame containing all possible modeled variables, such as from the MOM6 model. Must contain columns \code{Long.Name} and \code{Short.Name}, containing the full names and abbreviated names of the variables. Abbreviated names should correspond to those in the variable importance plots}

\item{template}{a character string designating the .qmd file to use as a template}

\item{report_path}{character string designating where the reports should be saved}
}
\value{
Function does not return anything. Reports are saved to \code{report_path}
}
\description{
Builds SDM report using a provided template. Requires directory to be set up per directions in the package documentation/manual.
}
