% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_variable_exposure.R
\name{make_variable_exposure}
\alias{make_variable_exposure}
\title{Make Maps or Timeseries of Variable-Specific Exposure}
\usage{
make_variable_exposure(type, ranked_exposure, sdm_raster)
}
\arguments{
\item{type}{designates desired output, must equal 'map' or 'timeseries'}

\item{ranked_exposure}{output from \code{rank_exposure}}

\item{sdm_raster}{rasterStack with monthly averages of SDM results to use as weights for weighted average}
}
\value{
If \code{type == 'map'}, a rasterStack with the number of layers equal to the number of variables supplied, containing the weighted average exposure, weighted by SDM results, for each variable. If \code{type == 'timeseries'}, a matrix with the number of rows equal to the number of variables supplied, and 12 columns (1 for each month) containing the weighted average exposure, weighted by SDM results, for each variable.
}
\description{
Combines raw variable exposures and SDM results, using the SDM results as weights to average the exposure across time to produce maps of species-specific exposure for each variable or across space to produce timeseries of species-specific exposure for each variable
}
