% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_model_data.R
\name{normalize_model_data}
\alias{normalize_model_data}
\title{Normalize Model Data}
\usage{
normalize_model_data(raw_list, avg_list, sd_list, short_names)
}
\arguments{
\item{raw_list}{List of output rasters from \code{pull_hind} or \code{pull_forecast}. Alternatively, a list of rasters containing gridded data to be normalized, with each entry in the list containing a rasterStack of a timeseries of environmental data to be normalized.}

\item{avg_list, sd_list}{List of output rasters from \code{avg_model_data} and \code{sd_env}, respectively. Alternatively, averages and standard deviations of desired gridded datasets. Should be a list of rasterStacks, with each entry in the list containing the rasterStack for a different variable.}

\item{short_names}{vector of simplified variable names to help name resulting raster files.}
}
\value{
a list whose length is equal to the length of \code{raw_list}, where each item in the list is a rasterStack of normalized data associated with that variable
}
\description{
Normalize model data across the entire provided timeseries using a z-score
}
