% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_mom6_hindcast.R
\name{pull_mom6_hindcast}
\alias{pull_mom6_hindcast}
\title{Pull MOM6 Hindcast Data}
\usage{
pull_mom6_hindcast(
  var_url,
  req_var,
  gt = "regrid",
  of = "monthly",
  bounds = c(-78, -65, 35, 45),
  release
)
}
\arguments{
\item{var_url}{URL pointing to JSON table variable lists for desired MOM6 hindcast and domain}

\item{req_var}{name of variable to pull. Must match names in the 'cefi_long_name' column provided JSON table}

\item{gt}{desired grid type. Must match one of the options in the 'cefi_grid_type' column in provided JSON table}

\item{of}{desired output frequency. Must match one of the options in the 'cefi_output_frequency' column in provided JSON table}

\item{bounds}{xmin, xmax, ymin, ymax of desired output raster}

\item{release}{release code. Must match one of the options in the 'cefi_release' column in provided JSON table}
}
\value{
a spatRaster of data associated with the requested variable
}
\description{
Pull hindcast data from the MOM6 model output based on provided URL from the CEFI portal
}
