% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_corr.R
\name{remove_corr}
\alias{remove_corr}
\title{Remove Correlated Environmental Covariates from Species/Environmental Data Frame}
\usage{
remove_corr(se, pa_col, xy_col, month_col, year_col)
}
\arguments{
\item{se}{data.frame of fisheries and environmental data}

\item{pa_col}{column name for presence/absence column}

\item{xy_col}{a vector with a length of 2 indicating the longitude and latitude column names}

\item{month_col, year_col}{column names for month and year columns respectively}
}
\value{
a data frame with correlated covariates removed
}
\description{
This function removed correlated covariates to prepare data for modeling
}
